requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
asgiref==3.5.2
astor==0.8.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
blinker==1.5
blis==0.7.9
cachetools==5.2.0
catalogue==2.0.8
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
chardet==5.0.0
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.5
commonmark==0.9.1
confection==0.0.4
cycler==0.11.0
cymem==2.0.7
DataProperty==0.55.0
datasets==2.6.1
decorator==5.1.1
dill==0.3.5.1
Django==4.1.2
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
executing==1.0.0
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
huggingface-hub==0.9.1
idna==3.4
image==1.5.33
importlib-metadata==4.12.0
importlib-resources==5.10.0
inflect==6.0.2
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==1.5.2
jsonschema==4.17.0
kiwisolver==1.4.4
langcodes==3.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mbstrdecoder==1.1.1
multidict==6.0.2
multiprocess==0.70.13
munch==2.5.0
murmurhash==1.0.9
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.5.3.56
overrides==6.2.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathtools==0.1.2
pathvalidate==2.5.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
preshed==3.0.8
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==8.0.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydantic==1.10.2
pydeck==0.8.0
pyDeprecate==0.3.2
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
pytablewriter==0.64.2
python-dateutil==2.8.2
pytorch-lightning==1.6.5
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rsa==4.9
sacred==0.8.2
scikit-learn==1.1.3
scipy==1.9.1
semver==2.13.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
stack-data==0.5.0
streamlit==1.3.1
tabledata==1.3.0
tabulate==0.9.0
tcolorpy==0.1.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==8.1.9
threadpoolctl==3.1.0
timm==0.6.7
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==0.9.3
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.0
typepy==1.3.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wandb==0.13.3
wasabi==1.1.1
watchdog==2.1.9
wcwidth==0.2.5
Werkzeug==2.2.2
wordcloud==1.8.2.2
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.8.1
